+++ lib/Controller/SettingsController.php
@@ -136,6 +136,7 @@ class SettingsController extends Controller {
             "currentServer" => $this->urlGenerator->getAbsoluteURL("/"),
             "formats" => $this->config->formatsSetting(),
             "sameTab" => $this->config->getSameTab(),
+            "enableSharing" => $this->config->getEnableSharing(),
             "preview" => $this->config->getPreview(),
             "advanced" => $this->config->getAdvanced(),
             "cronChecker" => $this->config->getCronChecker(),
@@ -225,6 +226,7 @@ class SettingsController extends Controller {
      * @param array $defFormats - formats array with default action
      * @param array $editFormats - editable formats array
      * @param bool $sameTab - open in the same tab
+     * @param bool $enableSharing - enable sharingsameTab
      * @param bool $preview - generate preview files
      * @param bool $advanced - use advanced tab
      * @param bool $cronChecker - disable cron checker
@@ -246,6 +248,7 @@ class SettingsController extends Controller {
         $defFormats,
         $editFormats,
         $sameTab,
+        $enableSharing,
         $preview,
         $advanced,
         $cronChecker,
@@ -266,6 +269,7 @@ class SettingsController extends Controller {
         $this->config->setDefaultFormats($defFormats);
         $this->config->setEditableFormats($editFormats);
         $this->config->setSameTab($sameTab);
+        $this->config->setEnableSharing($enableSharing);
         $this->config->setPreview($preview);
         $this->config->setAdvanced($advanced);
         $this->config->setCronChecker($cronChecker);