+++ src/main.js
@@ -178,12 +178,27 @@ import { loadState } from '@nextcloud/initial-state'
 		if (winEditor && winEditor.location) {
 			OCA.Onlyoffice.SetDefaultUrl()
 			winEditor.location.href = url
-		} else if (!OCA.Onlyoffice.setting.sameTab || OCA.Onlyoffice.mobile || OCA.Onlyoffice.Desktop) {
+		} else if ((!OCA.Onlyoffice.setting.sameTab && !OCA.Onlyoffice.setting.enableSharing)
+			|| OCA.Onlyoffice.mobile || OCA.Onlyoffice.Desktop || (isPublicShare() && !OCA.Onlyoffice.isViewIsFile()
+			&& !OCA.Onlyoffice.setting.sameTab && OCA.Onlyoffice.setting.enableSharing)) {
 			OCA.Onlyoffice.SetDefaultUrl()
 			winEditor = window.open(url, '_blank')
 		} else if (isPublicShare() && OCA.Onlyoffice.isViewIsFile()) {
 			location.href = url
 		} else {
+			if (OCA.Onlyoffice.setting.enableSharing
+				&& !isPublicShare()
+				&& (window.OCP?.Files?.Router?.query?.openfile === undefined || window.OCP?.Files?.Router?.query?.openfile === 'false')) {
+				window.OCP?.Files?.Router?.goToRoute(
+					null, // use default route
+					{ view: 'files', fileid: fileId },
+					{ ...OCP.Files.Router.query, openfile: 'true' },
+				)
+				url = window.location.href
+				OCA.Onlyoffice.SetDefaultUrl()
+				window.open(url, '_blank')
+				return
+			}
 			OCA.Onlyoffice.frameSelector = '#onlyofficeFrame'
 			const $iframe = $('<iframe id="onlyofficeFrame" nonce="' + btoa(OC.requestToken) + '" scrolling="no" allowfullscreen src="' + url + '&inframe=true" />')
 
@@ -885,7 +900,6 @@ import { loadState } from '@nextcloud/initial-state'
 			OCA.Onlyoffice.registerAction()
 		}
 	}
-
 	initPage()
 
 })(OCA)