+++ src/settings.js
@@ -225,6 +225,7 @@
 			})
 
 			const sameTab = $('#onlyofficeSameTab').is(':checked')
+			const enableSharing = $('#onlyofficeEnableSharing').is(':checked')
 			const preview = $('#onlyofficePreview').is(':checked')
 			const advanced = $('#onlyofficeAdvanced').is(':checked')
 			const cronChecker = $('#onlyofficeCronChecker').is(':checked')
@@ -251,6 +252,7 @@
 					defFormats,
 					editFormats,
 					sameTab,
+					enableSharing,
 					preview,
 					advanced,
 					cronChecker,
@@ -428,6 +430,16 @@
 
 			location.href = downloadLink
 		})
+
+		const sameTabCheckbox = document.getElementById('onlyofficeSameTab')
+		const sharingBlock = document.getElementById('onlyofficeEnableSharingBlock')
+		const sharingCheckbox = document.getElementById('onlyofficeEnableSharing')
+
+		sameTabCheckbox.onclick = function() {
+			const isChecked = sameTabCheckbox.checked
+			sharingBlock.style.display = isChecked ? 'none' : 'block'
+			sharingCheckbox.checked = isChecked ? sharingCheckbox.checked : false
+		}
 	})
 
 })(jQuery, OC)