+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
+
+##
 ## Added
-- shardKey param to url for requests to Conversion API
+- sharing panel when editing in a new tab
+- shardKey param to url for requests to Docs
 
 ## 9.7.0
 ## Added